refactor(ui): MessageItem 类型和工具函数独立到 message/ 目录
从 MessageList.tsx（2075 行）提取纯类型和工具函数到独立文件：
- message/types.ts：MessageItem, MessageGroup, SystemMessageMeta
- message/utils.ts：16 个纯函数（formatMessageTime, getToolMessageMeta,
  summarizeToolGroup, groupMessages, resolveAssetUrl 等）
- message/index.ts：统一导出

MessageList.tsx 改为 import，从 2075 行减少到 ~1850 行。
保留在 MessageList.tsx 的：React 组件、JSX 渲染、hook、Fragment 类型。
